requirements.txt
astroid==2.0.4
bleach==2.1.3
certifi==2018.8.13
chardet==3.0.4
colorama==0.3.9
Django==2.1
django-admin-tools==0.8.1
django-apptemplates==1.4
django-autoslug==1.9.3
django-autoslug-iplweb==1.9.4
django-bootstrap4==0.0.7
django-colorfield==0.1.15
django-crispy-forms==1.7.2
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-fobi==0.13.5
django-formtools==2.1
django-nine==0.1.13
django-nonefield==0.3
easy-thumbnails==2.5
gunicorn==19.9.0
html5lib==1.0.1
httplib2==0.11.3
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauth2client==4.1.2
pbr==4.2.0
Pillow==5.2.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pylint==2.1.1
pytz==2018.5
requests==2.19.1
rsa==3.4.2
simplejson==3.16.0
six==1.11.0
stevedore==1.29.0
typed-ast==1.1.0
Unidecode==1.0.22
uritemplate==3.0.0
urllib3==1.23
virtualenv==16.0.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
vishap==0.1.5
webencodings==0.5.1
whitenoise==4.0
wrapt==1.10.11
yolk3k==0.9
